pyproject.toml
[tool.poetry]
name = "smart_chatter"
version = "0.1.0"
description = "SmartChatter"
license = "UNLICENSED"
authors = [
    "extrasad <carlosr1765@gmail.com>",
]
readme = "README.md"
repository = "https://github.com/extrasad/smart-chatter"
classifiers = [
    "Topic :: Internet :: WWW/HTTP :: WSGI :: ChatBot",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Natural Language :: English",
]
include = ["pyproject.toml"]

[tool.poetry.dependencies]
python = "^3.6.8"
django = ">=2.2.13"
gunicorn = "*"
django-heroku="*"
chatterbot = "*"
djangorestframework = "*"
chatterbot-corpus = {git = "https://github.com/gunthercox/chatterbot-corpus.git"}
graphene-django = "*"
spacy = "*"
pyyaml = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"